Moved standalone functions to utils
