+++ sharding/handler/log_handler.py
@@ -1,41 +1,9 @@
 import logging
 
-
-def get_recent_block_hashes(w3, history_size):
-    block = w3.eth.getBlock('latest')
-    recent_hashes = []
-
-    for _ in range(history_size):
-        recent_hashes.append(block['hash'])
-        # break the loop if we hit the genesis block.
-        if block['number'] == 0:
-            break
-        block = w3.eth.getBlock(block['parentHash'])
-
-    return tuple(reversed(recent_hashes))
-
-
-def get_canonical_chain(w3, recent_block_hashes, history_size):
-    block = w3.eth.getBlock('latest')
-
-    new_block_hashes = []
-
-    for _ in range(history_size):
-        if block['hash'] in recent_block_hashes:
-            break
-        new_block_hashes.append(block['hash'])
-        block = w3.eth.getBlock(block['parentHash'])
-    else:
-        raise Exception('No common ancestor found')
-
-    first_common_ancestor_idx = recent_block_hashes.index(block['hash'])
-
-    revoked_hashes = recent_block_hashes[first_common_ancestor_idx + 1:]
-
-    # reverse it to comply with the order of `self.recent_block_hashes`
-    reversed_new_block_hashes = tuple(reversed(new_block_hashes))
-
-    return revoked_hashes, reversed_new_block_hashes
+from sharding.handler.utils.web3_utils import (
+    get_recent_block_hashes,
+    get_canonical_chain,
+)
 
 
 class LogHandler:
